commenting

+++ congress/models.py
@@ -81,32 +81,41 @@ class CongressPerson(models.Model):
     termsServed = models.JSONField(default=None, blank=True, null=True)
     
     # total transactions
-    # Big Integer Field is a field that can store a large integer value (the regular integerfields can only hold )
-    totalTransactions = models.BigIntegerField(default=0)
+    # An integer field hold a integer of max value 2147483647
+    totalTransactions = models.IntegerField(default=0)
 
     # Total Volume of transactions
-    totalVolumeTransactions = models.BigIntegerField(blank=True, null=True, default=0)
+    totalVolumeTransactions = models.IntegerField(blank=True, null=True, default=0)
 
-    # Trade Type Ratio
-    purchases = models.BigIntegerField(blank=True, null=True, default=0)
-    sales = models.BigIntegerField(blank=True, null=True, default=0)
+    # purchase
+    purchases = models.IntegerField(blank=True, null=True, default=0)
+    # sales
+    sales = models.IntegerField(blank=True, null=True, default=0)
 
+    # String representation of object
+    # Example without function below: <QuerySet [<Blog:>,<Blog:>,<Blog:>....]
+    # Example with function below:  <QuerySet [<Blog:itsName>,<Blog:itsName>,<Blog:itsName>....]
     def __str__(self):
         return self.fullName
 
     # Update congress persons transaction count every time the object is saved 
     def updateStats(self):
+        # Query database for all transactions for this congress person
         transactions = CongressTrade.objects.filter(name=self)
 
         # update model
         CongressPerson.objects.filter(bioguide=self.bioguide).update(
-            totalTransactions=transactions.count(), 
+            # Count number of transactions
+            totalTransactions=transactions.count(),
             purchases=transactions.filter(transactionType='Purchase').count(), 
+            # Calculate the number of Sales
+            # Here we use "startswith" because we have two options: "Sale (Full)" and "Sale (Partial)". Both of those options start with "Sale"
             sales=transactions.filter(transactionType__startswith='Sale').count(),
+            # Calculate the total volume of transactions
             totalVolumeTransactions=getSumMid(transactions), 
         )
         
-# Signals to update total transactions for each congress member
+# Signals to update total transactions for each congress member when there has beed a save(post_save) or deletion(post_delete)
 signals.post_save.connect(congressPersonStatUpdate, sender=CongressPerson)
 signals.post_delete.connect(congressPersonStatUpdate, sender=CongressPerson)
 
@@ -119,6 +128,7 @@ class Ticker(models.Model):
     company = models.CharField(max_length=1000)
 
     # marketcap 
+    # Big Integer Field is a field that can store a large integer value (the regular integerfields can only hold max 2147483647, and we need to store bigger values for a stocks marketcap)
     marketcap = models.BigIntegerField(blank=True, null=True)
     # sector
     sector = models.CharField(max_length=1000)
@@ -127,30 +137,37 @@ class Ticker(models.Model):
 
     # Signals to update total transactions for each congress member
     # total transactions it occured in
-    totalTransactions = models.BigIntegerField(blank=True, null=True, default=0)
+    totalTransactions = models.IntegerField(blank=True, null=True, default=0)
 
     # Total Volume of transactions
-    totalVolumeTransactions = models.BigIntegerField(blank=True, null=True, default=0)
+    totalVolumeTransactions = models.IntegerField(blank=True, null=True, default=0)
 
     # Trade Type Ratio
-    purchases = models.BigIntegerField(blank=True, null=True, default=0)
-    sales = models.BigIntegerField(blank=True, null=True, default=0)
-
+    purchases = models.IntegerField(blank=True, null=True, default=0)
+    sales = models.IntegerField(blank=True, null=True, default=0)
     
+    # String representation of object
+    # Example without function below: <QuerySet [<Blog:>,<Blog:>,<Blog:>....]
+    # Example with function below:  <QuerySet [<Blog:itsName>,<Blog:itsName>,<Blog:itsName>....]
     def __str__(self):
         return self.ticker
 
-
     # Update congress persons transaction count every time the object is saved 
     def updateStats(self):
+        # Query database for all transactions for this stock ticker
         transactions = CongressTrade.objects.filter(ticker=self)
 
         # update model
         try:
             Ticker.objects.filter(ticker=self).update(
+                # Count number of transactions
                 totalTransactions=transactions.count(), 
+                # Calculate the number of purchases 
                 purchases=transactions.filter(transactionType='Purchase').count(), 
+                # Calculate the number of Sales
+                # Here we use "startswith" because we have two options: "Sale (Full)" and "Sale (Partial)". Both of those options start with "Sale"
                 sales=transactions.filter(transactionType__startswith='Sale').count(),
+                # Calculate the total volume of transactions
                 totalVolumeTransactions=getSumMid(transactions), 
             )
         except Exception as e:
@@ -162,16 +179,18 @@ signals.post_delete.connect(tickerStatUpdate, sender=Ticker)
     
 # Combination of Senator and House Data
 class CongressTrade(models.Model):
-    # PRIMARY KEY (connect to another table with congress persons names)
+    # Foreign KEY (connect to another table with congress persons names, see ERD Diagram in InsiderUnlocked/Insider-Unlocked for more information)
     name = models.ForeignKey(CongressPerson, on_delete=models.CASCADE)
     
-    # PRIMARY KEY (connect to another table with ticker name)
+    # Foreign KEY (connect to another table with ticker name, see ERD Diagram in InsiderUnlocked/Insider-Unlocked for more information)
     ticker = models.ForeignKey(Ticker, on_delete=models.CASCADE, null=True)
 
     # transaction date
+    # A DateField is a field that can store a date value
     transactionDate = models.DateField(blank=True, null=True)
     
     # disclosure date
+    # A DateField is a field that can store a date value
     disclosureDate = models.DateField(blank=True, null=True)
 
     # transaction type
@@ -199,13 +218,17 @@ class CongressTrade(models.Model):
     comment = models.CharField(max_length=1000, blank=True)
     
     # is the transaction a pdf
+    # A BooleanField is a field that can store a boolean value
     pdf = models.BooleanField()
     
-
+    # String representation of object
+    # Example without function below: <QuerySet [<Blog:>,<Blog:>,<Blog:>....]
+    # Example with function below:  <QuerySet [<Blog:itsName>,<Blog:itsName>,<Blog:itsName>....]
     def __str__(self):
         return self.name.fullName
     
-
+    # Class we can use in django to create a unique_together constraint (can only create fields if they have a different value that an existing one combined with the fields listed), and ordering (orders the trades done by congress by the transaction date)
+    # Class we can use in django to create a unique_together constraint (can only create fields if they have a different value that an existing one combined with the fields listed), and ordering (orders the trades done by congress by the transaction date)
     class Meta:
         unique_together = ('disclosureDate', 'transactionDate', 'owner', 'ticker', 'assetDescription', 'assetType', 'transactionType', 'amount', 'comment', 'name', 'pdf', 'ptrLink')
         ordering = ["-transactionDate"]
@@ -219,22 +242,31 @@ class SummaryStat(models.Model):
     totalVolume = models.BigIntegerField(default=0)
     timeframe = models.BigIntegerField(unique=True)
     
+    # String representation of object
+    # Example without function below: <QuerySet [<Blog:>,<Blog:>,<Blog:>....]
+    # Example with function below:  <QuerySet [<Blog:itsName>,<Blog:itsName>,<Blog:itsName>....]
     def __str__(self):
         return str(self.total)
             
+    # Class we can use in django to create a unique_together constraint (can only create fields if they have a different value that an existing one combined with the fields listed), and ordering (orders the trades done by congress by the transaction date)
     class Meta:
         unique_together = ('total', 'purchases', 'sales', 'totalVolume', 'timeframe',)
-        # ordering = ["-transactionDate"]
 
     # Update congress persons transaction count every time the object is saved 
     def updateStats(self):
+        # Query database for all transactions for this congress person
         transactions = CongressTrade.objects.filter(transactionDate__lte=datetime.datetime.today(), transactionDate__gt=datetime.datetime.today()-datetime.timedelta(days=self.timeframe))
 
         # update model
         SummaryStat.objects.filter(id=self.id).update(
+            # Count number of transactions
             total=transactions.count(), 
+            # Count number of transactions
             purchases=transactions.filter(transactionType='Purchase').count(), 
+            # Calculate the number of Sales
+            # Here we use "startswith" because we have two options: "Sale (Full)" and "Sale (Partial)". Both of those options start with "Sale"
             sales=transactions.filter(transactionType__startswith='Sale').count(),
+            # Calculate the total volume of transactions
             totalVolume=getSumMid(transactions), 
         )
